+++ data/lexicon/switches/person-date.js
@@ -2,6 +2,7 @@
 export default [
   // clues: [person, date],
   // fallback: 'Month',
-  'april', 'august', 'jan', 'january', 'june', 'sep', 'avril'
-]//may
+  'april', 'august', 'jan', 'january', 'june', 'sep', 'avril',
+  // 'may'
+]