+++ src/3-three/coreference/api.js
@@ -1,0 +1,7 @@
+const api = function (View) {
+  // 
+  View.prototype.references = function () {
+
+  }
+}
+export default api