+++ src/3-three/coreference/plugin.js
@@ -1,0 +1,7 @@
+import coreference from './compute/index.js'
+import api from './api.js'
+
+export default {
+  compute: { coreference },
+  api
+}